Drive Monitoring/Load Pins Web V1
On this version we added the mongodb docker compose file to run the mongodb container, we modified the ip on the mongo_utils to communicate to this container and we tried to create a Django container but it was not working locally. We fixed some animation errors and modified the settings adding the accepted hosts.